Moving some things into a code folder for better organization!
