+++ code/api/routes/items/delete.go
