+++ code/api/routes/items/list.go
