+++ code/api/routes/server/register.go
