+++ main.go
@@ -1,8 +1,8 @@
-package main 
+package main
 
 import (
-	"github.com/Voltaic314/GhostFS/api"
-	"github.com/Voltaic314/GhostFS/db/seed"
+	"github.com/Voltaic314/GhostFS/code/api"
+	"github.com/Voltaic314/GhostFS/code/db/seed"
 )
 
 func main() {
